requirements.txt
python-magic
pycatbox
requests
cloudscraper
